requirements.txt
absl-py==0.5.0
astor==0.7.1
gast==0.2.0
grpcio==1.15.0
Markdown==2.6.11
numpy==1.14.5
opencv-python==3.4.3.18
pandas==0.23.4
Pillow==5.2.0
pkg-resources==0.0.0
protobuf==3.6.1
python-dateutil==2.7.3
pytz==2018.5
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
Werkzeug==0.14.1
